requirements.txt
aiogram==3.10.0
gspread==6.1.0
oauth2client==4.1.3
python-dotenv==1.0.1
